[backend] Add attributes to fields if they are not present
Namely, `fieldClass` and `fieldHelp`. Since these attributes are
updated programatically on form validation (eg. `fieldClass` can be set
to `is-danger` or `is-success`), if they are not already present when
Vue.js handle the field schema, they won't be added afterward because of
the way Vue.js work with reactive properties.