Add decorator that add permissions on each feature
More precisely, when a RESTful view for `GET` verb return a GeoJSON feature collection,
this decorator can be used to add a `can_edit` property (true or false) to each feature in the
collection.

`can_edit` is true or false depending on the current user role in the feature's study.